[qgis] Add menu and dropdown. (#2634)
Fixes #2627
Fixes #2588

This removes the layout dock, and makes the Ribasim button the "open"
action. The same action is now present in the plugins menu. Furthermore,
you can now also drag and drop a ribasim toml file. Finally, this also
loads all chunks of an arrow result file, instead of only the first one.

This also significantly refactors the codebase:

- Dropping the Input nodes altogether
- Removed all references to editing
- Removed all references to layouts/docks

The codebase can be further improved, I already removed the inheritance
of QWidget in the DatasetWidget, and it could be further slimmed down to
standalone RibasimModel class. The RibasimWidget can then be further
refactored, and possibly storing references to multiple RibasimModels,
instead of the single DatasetWidget that is overwritten by a newly
opened model.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>