Go back to Julia 1.11 (#2649)
This reverts #2629 and #2644.

The main reason is that we were only able to build working binaries with
cpu_target native. That meant the TeamCity binaries were much less
portable to other machines than before. Once that is resolved (we made
an upstream issue in JuliaC), we can go back to 1.12.