Get QGIS plugin ready to publish (#2654)
In https://github.com/Deltares/Ribasim/issues/1632 there was some
discussion on when to start publishing the QGIS plugin. After the
decision to remove editing features, executed in
https://github.com/Deltares/Ribasim/pull/2472, and subsequent UX
improvements in https://github.com/Deltares/Ribasim/pull/2634, I feel
like we have a pretty decent plugin that is ready for a first published
release. Installing from the plugin manager is easier for users, and
with the DSD and pizza courses coming up I think now is a good time.

My plan was to release Ribasim v2025.6.0 in a week or so. Besides this
PR I don't expect more developments for QGIS in that week. So here I
bump the versions for the plugin only, such that we can get the release
process underway. I'm not sure how long that takes because there is some
manual review.

Now that the black screen in the iMOD timeseries widget is fixed by
https://github.com/Deltares/imod-qgis/pull/89 and released as iMOD QGIS
v0.5.3, we can remove the temporary override we had in place and
document this as the minimum required version. We could also keep it
longer, but since the override is not really proper behavior I'd rather
remove it before the review.

~~EDIT: One more thing I'd like to change before releasing is to not
show all results, but only flow, by default.~~ (done)