Formulate in terms of flows, solve in terms of storages (#2624)
TL;DR: Computing the Jacobian with the performance of storage states
without losing the water balance accuracy of flow states.

Our state vector $u$ consists mostly of cumulative flows, which we sum
in our rhs to obtain the storages. Hence our rhs is of the form $f(Au)$,
where $A$ is a highly sparse matrix, and $Au$ gives the sums of all the
cumulative flows without other storage contributions like those from
flow boundary and initial storage (and the PID integral terms). Using
the chain rule, the Jacobian w.r.t. $u$ is

$$
   J_u \[f(Au)](u) = J\[f](Au)A,
$$ 

i.e. the Jacobian of $f$ in $Au$ multiplied by $A$. In the code $Au$ is
denoted by `u_reduced`. The Jacobian that is now computed with automatic
differentiation is only $J\[f](u_\text{reduced})$, which is a matrix
with the same number of rows as the full Jacobian but much fewer colums.
Furthermore, the product $Au$ is computed much less frequently because
it only happens outside rhs calls.

It turns out that this is very beneficial for the lhm, making the
Jacobian evaluation so much cheaper that it went from >80% of the
runtime to <10%.

Since the code around AD is becoming somewhat elaborate and complex, I
introduced a new file `differentiation.jl`.

(The branch name is not accurate)

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>