Improved testing for subnetwork communication (#2647)
Continues on #2556 and fixes #2632.

The allocation between subnetworks and the primary network had was
developed in #2556 and tested on a minimal model with a single secondary
network and 1 route for the flow.

In this PR, a more complex test network was created and showed several
bugs which were fixed.
In the model below allocation was calculated once as a single network
and also as a primary network connected to two secondary networks.

The inflow from the flow boundary of 0.9 m3/s was exactly matched by the
3 user demands of 0.3 m3/s. The results of allocation were compared with
the single network version and matched very well (atol = 1e-8).


<img width="1022" height="679" alt="Screenshot 2025-10-24 at 13 44 25"
src="https://github.com/user-attachments/assets/32e50840-8e84-4744-be9e-07545a0b62a2"
/>