Remove hardcoded or even magic constants. (#2592)
It seems good to document hardcoded or magic constants, and makes them
configurable in the toml.

This still requires better naming (and maybe better defaults?). Feel
free to take over this PR when you know some.

Based on a small discussion with @DanielTollenaar (odd behaviour when
trying to maintain levels within cm ranges).

---------

Co-authored-by: Lex <90280242+simulutions@users.noreply.github.com>