Add ResidenceTime tracer (#2701)
One of my shorter PRs, quite effective I'd say.

This adds a ResidenceTime tracer, which tracks the average *age* of the
water. Since we assume instantaneous perfect mixing (we're tracking
water after all), this equals the average residence time. You could also
track it using `storage/outflow_rate`, but that uses average flows, and
can't handle no (Inf) or negative outflow.

The PR is short because age is easy to calculate; all inflows have zero
concentration (age) by default, and we add a simple aging method (+dt)
for all mass already in the basin. In contrast, tracking age for a
tracer is much more involved.

HWS with the tracer visualised in days (output is seconds) at the end of
the simulation, nicely showing dead zones in the model.
<img width="664" height="745" alt="Screenshot 2025-11-09 at 21 21 23"
src="https://github.com/user-attachments/assets/f97205f9-a652-4a95-bdec-f3b959c2c940"
/>

In preparation of #2438. Given the `dt`s of hours in the HWS model
above, one can see that tracers along the main rivers will be delayed,
as their residence times have a similar order. If you would divide this
output by the `dt` you have the Courant number (or as close to it in a
0d model like this).