Fix the last Heisenbug (#2691)
Fixes #2670  and #2639


Summary: 
The last allocation heisenbug was caused due to reading uninitialized
memory during the jacobian generation.

Uninitialized memory cause:
The problem is that we would store a constant scalar flow_rate for pump
or outlet as a time dependent interpolations in the
'StateAndTimeDependentCache'. For optimization purpose, we minimize the
amount of interpolation evaluations since these are expensive.

What we do is check if the flow rate has changed, or if we are at a new
time step, since only then would we need to evaluate the
time-interpolation to get the new flow rate.

However, based on the constant flow_rate, this is never the case, which
gave us problems at t = 0.

The solution was as simple as to just have different caches for time
independent flow rates (constant scalars) and time dependent flow rates
(interpolation objects).

---------

Co-authored-by: Maarten Pronk <git@evetion.nl>