Polder management with allocation (#2739)
Improves #2723.

In the communication between the physical and allocation layer, the
vertical forcing lagged behind 1 full allocation time step. The vertical
foring was integrated and averaged over the last physical timestep, and
then assumed constant for the next allocation time step. This resulted
in lagging pump behaviour which looked unrealistic.

What we do now in allocation is: 

1. look up at the beginning of the allocation time step what the
vertical forcings are,
2. assume them constant over the timestep, just as the area of the basin
at the level on the beginning of the time step
3. and compute an euler backward step in the optimization problem.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>