[Delwaq] More includes for LWKM. (#2745)
This makes it easier to edit dry loads (e.g. mass injections) as used by
LWKM (the other input can be stored in a Ribasim schematization). I've
made some small changes to better distinguish what's written by Ribasim
and what can be edited.

Added a version number so it's easier to check what version of the
generate script was used.

@gpijcke This partially solves
https://github.com/Deltares/Ribasim-NL/issues/392.