Add indexes to geopackage. (#2763)
This adds an index to each node_id (and link_id) column in the
geopackage. This makes identification in QGIS much faster, and even the
LHM now feels snappy (before it took a few seconds for a node to be
identified).

The root cause is that the (QGIS specific) relations between node tables
required a lot of lookups, and without an index, that's a full search in
tables. That becomes noticeable at the scale of the LHM. This PR does
make geopackages 10% larger.

Ideally we would also make the relations explicit in the geopackage, but
(geo)pandas doesn't support that. Creating tables ourselves is doable,
but not sure if it's worth it.

Also fixes #1973, by not opening the database (triggered by setting the
filepath) while writing it.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>