Bump version to 2026.1.0-dev1
Version numbers in the docs will still point to the latest release:

- docs.toml: document latest release
- install.ps1: install latest release
- install.qmd: document latest release (add pre-release section later)

AppxManifest.toml is not currently used, but I set it to `2026.1.0.0` because we cannot put a pre-release name there. Possibly that means we need to keep counting and use e.g. `2026.1.0.4` for the first stable release if we have 3 pre-releases, if we want to support upgrading from pre-release versions.