Bump version to 2026.1.0-rc1 (#2787)
Version numbers in the docs will still point to the latest release:

- docs.toml: document latest release
- install.ps1: install latest release
- install.qmd: document latest release (add pre-release section later)

AppxManifest.toml is not currently used, but I set it to `2026.1.0.1`
because we cannot put a pre-release name there. We could reserve 1-99
for rcs, and start releases at 100.


This is the first pre-release running up to the stable v2026.1.0
release.
It is a release candidate, and we recommend most users to stick with
v2025.6.0 for now.
We expect to make more breaking changes and add features before the
final release.

(Often release candidates are used when no breaking changes or features
will be added, but for simplicity the architects decided to stick only
with rc labels. I suggest we mainly refer to it as a pre-release.)